Split some question score components into separate files
